+++ src/LettrServiceProvider.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Lettr\Laravel;
 
+use Illuminate\Mail\Mailer;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\ServiceProvider;
 use Lettr\Laravel\Exceptions\ApiKeyIsMissing;
+use Lettr\Laravel\Mail\LettrPendingMail;
 use Lettr\Laravel\Transport\LettrTransportFactory;
 use Lettr\Lettr;
 
@@ -15,7 +17,7 @@ class LettrServiceProvider extends ServiceProvider
     /**
      * The package version.
      */
-    public const VERSION = '0.1.0';
+    public const VERSION = '0.2.0';
 
     /**
      * Bootstrap any application services.
@@ -29,6 +31,21 @@ class LettrServiceProvider extends ServiceProvider
             $lettr = $this->app->make('lettr');
 
             return new LettrTransportFactory($lettr, $config['options'] ?? []);
+        });
+
+        Mailer::macro('lettr', function (): LettrPendingMail {
+            $facadeRoot = Mail::getFacadeRoot();
+
+            // When Mail::fake() is used, the facade root is already a Mailer (MailFake)
+            // Otherwise, use $this which is the Mailer instance the macro is called on
+            if ($facadeRoot instanceof \Illuminate\Contracts\Mail\Mailer) {
+                return new LettrPendingMail($facadeRoot);
+            }
+
+            /** @var Mailer $self */
+            $self = $this;
+
+            return new LettrPendingMail($self);
         });
     }